Sharable course schedule links sub-branch changes (#1492)
