+++ apps/antalmanac/src/components/RightPane/AddedCourses/AddedCoursePane.tsx
@@ -1,5 +1,5 @@
 import { Box, Chip, Paper, SxProps, TextField, Tooltip, Typography } from '@mui/material';
-import { AACourse } from '@packages/antalmanac-types';
+import { AACourse, ShortCourseSchedule } from '@packages/antalmanac-types';
 import { usePostHog } from 'posthog-js/react';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
@@ -11,9 +11,9 @@ import { ColumnToggleDropdown } from '$components/RightPane/CoursePane/CoursePan
 import SectionTableLazyWrapper from '$components/RightPane/SectionTable/SectionTableLazyWrapper';
 import { ClearScheduleButton } from '$components/buttons/Clear';
 import { CopyScheduleButton } from '$components/buttons/Copy';
+import { useIsReadonlyView } from '$hooks/useIsReadonlyView';
 import analyticsEnum, { logAnalytics } from '$lib/analytics/analytics';
 import { clickToCopy } from '$lib/helpers';
-import { useIsSharedSchedulePage } from '$src/hooks/useIsSharedSchedulePage';
 import AppStore from '$stores/AppStore';
 
 /**
@@ -238,7 +238,7 @@ function ScheduleNoteBox() {
 }
 
 function SkeletonSchedule() {
-    const getScheduleData = () => {
+    const getScheduleData = (): ShortCourseSchedule => {
         const skeletonSchedule = AppStore.getCurrentSkeletonSchedule();
         if (!skeletonSchedule.courses || skeletonSchedule.courses.length === 0) {
             const regularCourses = AppStore.schedule.getCurrentCourses();
@@ -252,6 +252,7 @@ function SkeletonSchedule() {
                     })),
                     customEvents: AppStore.schedule.getCurrentCustomEvents(),
                     scheduleNote: AppStore.getCurrentScheduleNote(),
+                    id: undefined,
                 };
             }
         }
@@ -336,7 +337,7 @@ function SkeletonSchedule() {
 }
 
 function AddedSectionsGrid() {
-    const isSharedSchedulePage = useIsSharedSchedulePage();
+    const isReadonlyView = useIsReadonlyView();
     const [courses, setCourses] = useState(getCourses());
     const [scheduleNames, setScheduleNames] = useState(AppStore.getScheduleNames());
     const [scheduleIndex, setScheduleIndex] = useState(AppStore.getCurrentScheduleIndex());
@@ -395,7 +396,7 @@ function AddedSectionsGrid() {
     return (
         <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
             <Box sx={{ display: 'flex', width: 'fit-content', position: 'absolute', zIndex: 2 }}>
-                {!isSharedSchedulePage && (
+                {!isReadonlyView && (
                     <>
                         <CopyScheduleButton index={scheduleIndex} buttonSx={buttonSx} />
                         <ClearScheduleButton buttonSx={buttonSx} analyticsCategory={analyticsEnum.addedClasses} />
