+++ apps/antalmanac/src/components/RightPane/AddedCourses/CustomEventDetailView.tsx
@@ -8,9 +8,9 @@ import { deleteCustomEvent } from '$actions/AppStoreActions';
 import { CustomEventDialog } from '$components/Calendar/Toolbar/CustomEventDialog/CustomEventDialog';
 import ColorPicker from '$components/ColorPicker';
 import { MapLink } from '$components/buttons/MapLink';
+import { useIsReadonlyView } from '$hooks/useIsReadonlyView';
 import analyticsEnum from '$lib/analytics/analytics';
 import buildingCatalogue from '$lib/locations/buildingCatalogue';
-import { useIsSharedSchedulePage } from '$src/hooks/useIsSharedSchedulePage';
 import AppStore from '$stores/AppStore';
 import { useTimeFormatStore } from '$stores/SettingsStore';
 
@@ -22,7 +22,7 @@ interface CustomEventDetailViewProps {
 const CustomEventDetailView = (props: CustomEventDetailViewProps) => {
     const { customEvent } = props;
     const { isMilitaryTime } = useTimeFormatStore();
-    const isSharedSchedulePage = useIsSharedSchedulePage();
+    const isReadonlyView = useIsReadonlyView();
 
     const [skeletonMode, setSkeletonMode] = useState(AppStore.getSkeletonMode());
 
@@ -103,7 +103,7 @@ const CustomEventDetailView = (props: CustomEventDetailViewProps) => {
                                 deleteCustomEvent(customEvent.customEventID, [AppStore.getCurrentScheduleIndex()]);
                             }}
                             size="large"
-                            disabled={isSharedSchedulePage}
+                            disabled={isReadonlyView}
                         >
                             <Delete fontSize="small" />
                         </IconButton>
