+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/ActionCell.tsx
@@ -8,8 +8,8 @@ import { addCourse, deleteCourse, openSnackbar } from '$actions/AppStoreActions'
 import ColorPicker from '$components/ColorPicker';
 import { TableBodyCellContainer } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/TableBodyCellContainer';
 import { useIsMobile } from '$hooks/useIsMobile';
+import { useIsReadonlyView } from '$hooks/useIsReadonlyView';
 import analyticsEnum, { logAnalytics } from '$lib/analytics/analytics';
-import { useIsSharedSchedulePage } from '$src/hooks/useIsSharedSchedulePage';
 import AppStore from '$stores/AppStore';
 
 /**
@@ -47,7 +47,7 @@ interface ActionProps {
  */
 export function ColorAndDelete({ section, term }: ActionProps) {
     const isMobile = useIsMobile();
-    const isSharedSchedulePage = useIsSharedSchedulePage();
+    const isReadonlyView = useIsReadonlyView();
 
     const flexDirection = isMobile ? 'column' : undefined;
 
@@ -71,7 +71,7 @@ export function ColorAndDelete({ section, term }: ActionProps) {
                 alignItems: 'center',
             }}
         >
-            {!isSharedSchedulePage && (
+            {!isReadonlyView && (
                 <IconButton onClick={handleClick}>
                     <Delete fontSize="small" />
                 </IconButton>
