+++ apps/antalmanac/src/routes/Home.tsx
@@ -1,25 +1,19 @@
 import { Stack } from '@mui/material';
-import { LocalizationProvider } from '@mui/x-date-pickers';
-import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFnsV2';
 import { useCallback, useEffect, useRef } from 'react';
 import Split from 'react-split';
 
 import { ScheduleCalendar } from '$components/Calendar/CalendarRoot';
-import { Header } from '$components/Header/Header';
-import { HelpMenu } from '$components/HelpMenu/HelpMenu';
-import InstallPWABanner from '$components/InstallPWABanner';
-import { NotificationSnackbar } from '$components/NotificationSnackbar';
-import PatchNotes from '$components/PatchNotes';
+import HomePageWrapper from '$components/HomePageWrapper';
 import { ScheduleManagement } from '$components/ScheduleManagement/ScheduleManagement';
 import { useIsMobile } from '$hooks/useIsMobile';
 import { BLUE } from '$src/globals';
 import { useScheduleManagementStore } from '$stores/ScheduleManagementStore';
 
-function MobileHome() {
+export function MobileHome() {
     return <ScheduleManagement />;
 }
 
-function DesktopHome() {
+export function DesktopHome() {
     const setScheduleManagementWidth = useScheduleManagementStore((state) => state.setScheduleManagementWidth);
 
     const scheduleManagementRef = useRef<HTMLDivElement>(null);
@@ -77,18 +71,5 @@ function DesktopHome() {
 export default function Home() {
     const isMobile = useIsMobile();
 
-    return (
-        <LocalizationProvider dateAdapter={AdapterDateFns}>
-            <PatchNotes />
-            <InstallPWABanner />
-
-            <Stack component="main" height="calc(100svh + env(safe-area-inset-top))">
-                <Header />
-                {isMobile ? <MobileHome /> : <DesktopHome />}
-            </Stack>
-
-            <NotificationSnackbar />
-            <HelpMenu />
-        </LocalizationProvider>
-    );
+    return <HomePageWrapper>{isMobile ? <MobileHome /> : <DesktopHome />}</HomePageWrapper>;
 }
