+++ apps/antalmanac/src/stores/Schedules.ts
@@ -11,10 +11,9 @@ import type {
 
 import { calendarizeCourseEvents, calendarizeCustomEvents, calendarizeFinals } from './calendarizeHelpers';
 
-import { getDefaultTerm } from '$lib/termData';
 import { getNextScheduleName } from '$lib/utils';
 import { WebSOC } from '$lib/websoc';
-import { getColorForNewSection } from '$stores/scheduleHelpers';
+import { createEmptySchedule, getColorForNewSection } from '$stores/scheduleHelpers';
 
 /**
  * Manages state of schedules. Only one instance is really needed for the app.
@@ -40,15 +39,7 @@ export class Schedules {
     constructor() {
         const scheduleNoteId = Math.random();
 
-        this.schedules = [
-            {
-                scheduleName: `${getDefaultTerm().shortName.replaceAll(' ', '-')}`,
-                courses: [],
-                customEvents: [],
-                scheduleNoteId: scheduleNoteId,
-                id: undefined,
-            },
-        ];
+        this.schedules = [createEmptySchedule(scheduleNoteId)];
         this.currentScheduleIndex = 0;
         this.previousStates = [];
         this.futureStates = [];
@@ -60,10 +51,6 @@ export class Schedules {
         const scheduleNames = this.getScheduleNames();
         scheduleNames.splice(scheduleIndex, 1);
         return getNextScheduleName(newScheduleName, new Set(scheduleNames));
-    }
-
-    getDefaultScheduleName() {
-        return getDefaultTerm().shortName.replaceAll(' ', '-');
     }
 
     /**
