+++ apps/antalmanac/src/stores/scheduleHelpers.ts
@@ -1,7 +1,7 @@
 import {
     amber,
-    brown,
     blue,
+    brown,
     cyan,
     deepOrange,
     deepPurple,
@@ -15,7 +15,9 @@ import {
     teal,
     yellow,
 } from '@mui/material/colors';
-import { ScheduleCourse } from '@packages/antalmanac-types';
+import { Schedule, ScheduleCourse, ShortCourseSchedule } from '@packages/antalmanac-types';
+
+import { getDefaultTerm } from '$lib/termData';
 
 const colorVariants: Record<string, string[]> = {
     blue: [blue[300], blue[200], blue[100], blue[400], blue[500]],
@@ -114,4 +116,28 @@ export function getColorForNewSection(newSection: ScheduleCourse, sectionsInSche
         defaultColors.find((materialColor) => !usedColors.includes(materialColor)) ||
         defaultColors[(defaultColors.indexOf(lastDefaultColor) + 1) % defaultColors.length]
     );
+}
+
+export function getDefaultScheduleName() {
+    return getDefaultTerm().shortName.replaceAll(' ', '-');
+}
+
+export function createEmptySchedule(scheduleNoteId: number): Schedule {
+    return {
+        scheduleName: getDefaultScheduleName(),
+        courses: [],
+        customEvents: [],
+        scheduleNoteId: scheduleNoteId,
+        id: undefined,
+    };
+}
+
+export function createEmptyShortCourseSchedule(): ShortCourseSchedule {
+    return {
+        scheduleName: getDefaultScheduleName(),
+        courses: [],
+        customEvents: [],
+        scheduleNote: '',
+        id: undefined,
+    };
 }
